consolidate tree functions in tree.py
