+++ sympy/rules/__init__.py
@@ -28,5 +28,5 @@ import traverse
 from rl import rm_id, unpack, flatten, sort, glom, distribute, rebuild
 from util import new
 from strat import (canon, condition, debug, typed, chain, null_safe, do_one,
-        exhaust, minimize, treeexec, greedyexec)
+        exhaust, minimize)
 import branch
