+++ sympy/rules/branch/__init__.py
@@ -1,3 +1,3 @@
 import traverse
 from strat import (canon, condition, debug, multiplex, exhaust, notempty,
-        chain, onaction, sfilter, yieldify, allexec)
+        chain, onaction, sfilter, yieldify)
