+++ sympy/rules/branch/strat.py
@@ -1,5 +1,5 @@
 from strat_pure import (exhaust, multiplex, debug, notempty, condition, chain,
-        onaction, sfilter, yieldify, allexec)
+        onaction, sfilter, yieldify)
 from traverse import top_down
 
 def canon(*rules):
