+++ sympy/rules/strat.py
@@ -1,5 +1,5 @@
 from strat_pure import (exhaust, memoize, condition, chain, debug, tryit,
-        null_safe, do_one, switch, minimize, treeexec, greedyexec)
+        null_safe, do_one, switch, minimize)
 from traverse import bottom_up, top_down
 
 def canon(*rules):
